modularize uploads gallery moda
